Remove all double

+++ src/cpu/ls_qr_householder.c
@@ -3,13 +3,13 @@
 #include <string.h>
 #include "gemm.c"
 
-double house(int m, double* x, double* v);
-double* houseHolderHelper(double beta, double* v, int m);
-void test_backwarderror(double* A, double* Q, double* R, int m, int n);
+float house(int m, float* x, float* v);
+float* houseHolderHelper(float beta, float* v, int m);
+void test_backwarderror(float* A, float* Q, float* R, int m, int n);
 
-double house(int m, double* x, double* v) {
-    double sigma = 0.0;
-    double beta = 0.0;
+float house(int m, float* x, float* v) {
+    float sigma = 0.0;
+    float beta = 0.0;
     for (int i = 1; i < m; i++) {
         sigma += x[i] * x[i];
     }
@@ -19,14 +19,14 @@ double house(int m, double* x, double* v) {
     } else if (sigma == 0 && x[0] < 0) {
         beta = -2.0;
     } else {
-        double mu = sqrt(x[0]*x[0] + sigma);
+        float mu = sqrt(x[0]*x[0] + sigma);
         if (x[0] <= 0) {
             v[0] = x[0] - mu;
         } else {
             v[0] = -sigma / (x[0] + mu);
         }
         beta = 2 * v[0] * v[0] / (sigma + v[0]*v[0]);
-        double v0 = v[0];
+        float v0 = v[0];
         if (v0 != 0) {
             for (int i = 0; i < m; i++) {
                 v[i] = v[i]/v0;
@@ -36,40 +36,40 @@ double house(int m, double* x, double* v) {
     return beta;
 }
 
-// double house(int m, double* x, double* v){
-//     double xNorm = 0;
+// float house(int m, float* x, float* v){
+//     float xNorm = 0;
 //     for (int i = 0; i < m; i++){
 //         xNorm += x[i] * x[i];
 //     }
 //     xNorm = sqrt(xNorm);
-//     double sign = (x[0] >= 0) ? 1 : -1;
-//     memcpy(v, x, m * sizeof(double));
+//     float sign = (x[0] >= 0) ? 1 : -1;
+//     memcpy(v, x, m * sizeof(float));
 //     v[0] = x[0] + sign * xNorm;
-//     double vtv = 0;
-//     double v0 = v[0];
+//     float vtv = 0;
+//     float v0 = v[0];
 //     for (int i = 0; i < m; i++){
 //         v[i] = v[i]/v0;
 //         vtv += v[i] * v[i];
 //     }
-//     double beta = 2 / vtv;
+//     float beta = 2 / vtv;
 //     return beta;
 // }
 
-void houseHolderQR(double* A, int m, int n){
+void houseHolderQR(float* A, int m, int n){
     for (int j = 0; j < n; j ++){
-        double *v = (double*)calloc((m-j), sizeof(double));
-        double *x = (double*)calloc((m-j), sizeof(double));
+        float *v = (float*)calloc((m-j), sizeof(float));
+        float *x = (float*)calloc((m-j), sizeof(float));
         for (int i = j; i < m; i++){
             v[i-j] = A[i*n+j];
         }
         for (int i = j; i < m; i++){
             x[i-j] = A[i*n+j];
         }
-        double beta = house(m-j, x, v);
-        double* H = houseHolderHelper(beta, v, m-j);
+        float beta = house(m-j, x, v);
+        float* H = houseHolderHelper(beta, v, m-j);
         // multiply by A submatrix which is m-j * n-j
-        double *A_sub = (double*)calloc((m-j)*(n-j), sizeof(double));
-        double *A_sub_updated = (double*)calloc((m-j)*(n-j), sizeof(double));
+        float *A_sub = (float*)calloc((m-j)*(n-j), sizeof(float));
+        float *A_sub_updated = (float*)calloc((m-j)*(n-j), sizeof(float));
         for (int i = j; i < m; i++){
             for (int k = j; k < n; k++){
                 A_sub[(i-j)*(n-j) + k-j] = A[i*n+k];
@@ -95,13 +95,13 @@ void houseHolderQR(double* A, int m, int n){
 }
 
 // generate I - beta * v * vT
-double* houseHolderHelper(double beta, double* v, int m){
-    double* result = (double*)calloc(m*m, sizeof(double));
-    double* identityMatrix = (double*)calloc(m*m, sizeof(double));
+float* houseHolderHelper(float beta, float* v, int m){
+    float* result = (float*)calloc(m*m, sizeof(float));
+    float* identityMatrix = (float*)calloc(m*m, sizeof(float));
     for (int i = 0; i < m; i++) {
         identityMatrix[i*m+i] = 1.0;
     }
-    double* vvT = (double*)calloc(m*m, sizeof(double));
+    float* vvT = (float*)calloc(m*m, sizeof(float));
     gemm(vvT, v, v, m, m, 1);
     for (int i = 0; i < m; i++){
         for (int j = 0; j < m; j++){
@@ -113,14 +113,14 @@ double* houseHolderHelper(double beta, double* v, int m){
     return result;
 }
 
-double* backSubstitution(double* R, double* y, int m, int n){
+float* backSubstitution(float* R, float* y, int m, int n){
     if (m < n) {
         fprintf(stderr, "Invalid matrix dimensions: m must be >= n\n");
         exit(EXIT_FAILURE);
     }
-    double* x = (double*)calloc(n, sizeof(double));
+    float* x = (float*)calloc(n, sizeof(float));
     for (int i = n-1; i >= 0; i--){
-        double sum = 0;
+        float sum = 0;
         for (int j = i+1; j < n; j++){
             sum += R[i*n+j] * x[j];
         }
@@ -134,25 +134,25 @@ double* backSubstitution(double* R, double* y, int m, int n){
     return x;
 }
 
-void houseHolderQRb(double* A, double* b, int m, int n) {
+void houseHolderQRb(float* A, float* b, int m, int n) {
     for (int j = 0; j < n; j ++){
-        double *v = (double*)calloc((m-j), sizeof(double));
-        double *x = (double*)calloc((m-j), sizeof(double));
+        float *v = (float*)calloc((m-j), sizeof(float));
+        float *x = (float*)calloc((m-j), sizeof(float));
         for (int i = j; i < m; i++){
             v[i-j] = A[i*n+j];
         }
         for (int i = j; i < m; i++){
             x[i-j] = A[i*n+j];
         }
-        double beta = house(m-j, x, v);
-        double* H = houseHolderHelper(beta, v, m-j);
-        double* b_sub = (double*) calloc(m-j, sizeof(double));
-        double* b_sub_updated = (double*) calloc((m-j), sizeof(double));
+        float beta = house(m-j, x, v);
+        float* H = houseHolderHelper(beta, v, m-j);
+        float* b_sub = (float*) calloc(m-j, sizeof(float));
+        float* b_sub_updated = (float*) calloc((m-j), sizeof(float));
         for (int i = 0; i < m-j; i++){
             b_sub[i] = b[i+j];
         }
         // vt*b
-        double vtb = 0.0;
+        float vtb = 0.0;
         for (int i = j; i < m; i++) {
             vtb += v[i-j] * b[i];
         }
@@ -164,8 +164,8 @@ void houseHolderQRb(double* A, double* b, int m, int n) {
         //     b[i] = b_sub_updated[i-j];
         // } 
         // multiply by A submatrix which is m-j * n-j
-        double *A_sub = (double*)calloc((m-j)*(n-j), sizeof(double));
-        double *A_sub_updated = (double*)calloc((m-j)*(n-j), sizeof(double));
+        float *A_sub = (float*)calloc((m-j)*(n-j), sizeof(float));
+        float *A_sub_updated = (float*)calloc((m-j)*(n-j), sizeof(float));
         for (int i = j; i < m; i++){
             for (int k = j; k < n; k++){
                 A_sub[(i-j)*(n-j) + k-j] = A[i*n+k];
@@ -197,20 +197,20 @@ void houseHolderQRb(double* A, double* b, int m, int n) {
         }
     }
 
-    // double* householderQRLS(double* A, double* b, int m, int n){
+    // float* householderQRLS(float* A, float* b, int m, int n){
     //     houseHolderQR(A, m, n);
     //     for (int j = 0; j < n; j++) {
-    //         double *v = malloc(sizeof(double) * (m-j));
+    //         float *v = malloc(sizeof(float) * (m-j));
     //         v[0] = 1.0;
     //         for (int k = j+1; k < m; k++) {
     //             v[k-j] = A[k*n+j];
     //         }
-    //         double vtv = 0.0;
+    //         float vtv = 0.0;
     //         for (int i = 0; i < m-j; i++) {
     //             vtv += v[i] * v[i];
     //         }
-    //         double beta = 2.0/vtv;
-    //         double vtb = 0.0;
+    //         float beta = 2.0/vtv;
+    //         float vtb = 0.0;
     //         for (int i = j; i < m; i++) {
     //             vtb += v[i-j] * b[i];
     //         }
@@ -222,7 +222,7 @@ void houseHolderQRb(double* A, double* b, int m, int n) {
     // }
 
 
-    double* householderQRLS(double* A, double* b, int m, int n){
+    float* householderQRLS(float* A, float* b, int m, int n){
         houseHolderQRb(A, b, m, n);
         // now b is updated to QTb and A is updated to R
         // solve Rx = Q^Tb
