+++ test/cpu/test_solver.c
@@ -1,10 +1,10 @@
     void test_house() {
         // set test cases here
         int m = 3;
-        double x[] = {2, 2, 1};
-        double v[] = {0, 0, 0};
+        float x[] = {2, 2, 1};
+        float v[] = {0, 0, 0};
         // 
-        double beta = house(m, x, v);
+        float beta = house(m, x, v);
         for (int i = 0; i < m; i++) {
         printf("x: %f, v: %f\n", x[i], v[i]);
         }
@@ -14,13 +14,13 @@
     void test_houseHolderHelper() {
         // set test case here
         int m = 3;
-        double x[] = {2, 2, 1};
-        // double v[] = {5, 2, 1};
-        // double beta = 2.0 / 30.0;
-        double v[] = {1, -2, -1};
-        double beta = 1.0 / 3.0;
+        float x[] = {2, 2, 1};
+        // float v[] = {5, 2, 1};
+        // float beta = 2.0 / 30.0;
+        float v[] = {1, -2, -1};
+        float beta = 1.0 / 3.0;
         // 
-        double* result = houseHolderHelper(beta, v, m);
+        float* result = houseHolderHelper(beta, v, m);
 
         printf("H:\n");
         for (int i = 0; i < m; i++){
@@ -29,7 +29,7 @@
         }
         printf("\n");
         }
-        double *Hx = malloc(sizeof(double) * m);
+        float *Hx = malloc(sizeof(float) * m);
         gemm(Hx, result, x, m, 1, m);
         printf("Hx: \n");
         for (int i = 0; i < m; i++){
@@ -43,13 +43,13 @@
         // set test case here
         // int m = 4;
         // int n = 3;
-        // double A[] = {1.0, -1.0, 4.0, 1.0, 4.0, -2.0, 1.0, 4.0, 2.0, 1.0, -1.0, 0.0};
+        // float A[] = {1.0, -1.0, 4.0, 1.0, 4.0, -2.0, 1.0, 4.0, 2.0, 1.0, -1.0, 0.0};
         // int m = 3;
         // int n = 3;
-        // double A[] = {2.0, -2.0, 18.0, 2.0, 1.0, 0.0, 1.0, 2.0, 0.0};
+        // float A[] = {2.0, -2.0, 18.0, 2.0, 1.0, 0.0, 1.0, 2.0, 0.0};
         int m = 4;
         int n = 3;
-        double A[] = {1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 10.0, 10.0, 11.0, 13.0};
+        float A[] = {1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 10.0, 10.0, 11.0, 13.0};
         printf("A:\n");
         for (int i = 0; i < m; i++){
         for (int j = 0; j < n; j++){
@@ -70,10 +70,10 @@
     void test_backSubstitution() {
         int m = 4;
         int n = 3;
-        double A[] = {1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 10.0, 10.0, 11.0, 13.0};
-        double b[] = {14.0, 32.0, 50.0, 68.0};
-        double R[] = {12.884099 , 14.591630 , 16.299161 , 0.0, 1.041315 , 2.082630 , 0.0, 0.0, 0.0, 0.0, 0.0, 0.0};
-        double y[] = {90.964842, 8.330522 , 0.000000 , 0.000000 };
+        float A[] = {1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 10.0, 10.0, 11.0, 13.0};
+        float b[] = {14.0, 32.0, 50.0, 68.0};
+        float R[] = {12.884099 , 14.591630 , 16.299161 , 0.0, 1.041315 , 2.082630 , 0.0, 0.0, 0.0, 0.0, 0.0, 0.0};
+        float y[] = {90.964842, 8.330522 , 0.000000 , 0.000000 };
         householderQRLS(A, b, m, n);
         printf("A back: \n");
         for (int i = 0; i < m; i++){
@@ -86,7 +86,7 @@
         for (int i = 0; i < m; i++){
             printf("%f ", b[i]);
         }
-        double *x = backSubstitution(A, b, m, n);
+        float *x = backSubstitution(A, b, m, n);
         printf("x: \n");
         for (int i = 0; i < n; i++) {
         printf("%f ", x[i]);
@@ -97,11 +97,11 @@
     void test_houseHolderQRLS() {
         int m = 4;
         int n = 3;
-        double A[] = {1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 10.0, 10.0, 11.0, 13.0};
-        double b[] = {14.0, 32.0, 50.0, 68.0};
+        float A[] = {1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 10.0, 10.0, 11.0, 13.0};
+        float b[] = {14.0, 32.0, 50.0, 68.0};
 
-        double *x = (double*) malloc(sizeof(double) * n);
-        memset(x, 0, sizeof(double) * n);
+        float *x = (float*) malloc(sizeof(float) * n);
+        memset(x, 0, sizeof(float) * n);
         x = householderQRLS(A, b, m, n);
         printf("x: \n");
         for (int i = 0; i < n; i++) {
@@ -110,7 +110,7 @@
         printf("\n");
     }
 
-    void test_backwarderror(double* A, double* Q, double* R, int m, int n){
+    void test_backwarderror(float* A, float* Q, float* R, int m, int n){
         // print R 
         FILE *file = fopen("R_matrix.txt", "w");
         if (file == NULL) {
@@ -138,10 +138,10 @@
         }
         fclose(file);
         // print (A - qr) / A
-        double* qr = (double*)malloc(m*n*sizeof(double));
+        float* qr = (float*)malloc(m*n*sizeof(float));
         gemm(qr, Q, R, m, n, n);
-        double error = 0;
-        double residual = 0;
+        float error = 0;
+        float residual = 0;
         for (int i = 0; i < m; i++){
             for (int j = 0; j < n; j++){
                 residual += (A[i*n+j] - qr[i*n+j]) * (A[i*n+j] - qr[i*n+j]);
@@ -149,7 +149,7 @@
         }
         residual = sqrt(residual);
         printf("residual: %f\n", residual);
-        double normA = 0;
+        float normA = 0;
         for (int i = 0; i < m; i++){
             for (int j = 0; j < n; j++){
                 normA += A[i*n+j] * A[i*n+j];
